+++ memberaudit/__init__.py
@@ -1,4 +1,4 @@
 default_app_config = "memberaudit.apps.MemberauditConfig"
 
-__version__ = "1.0.0a2"
+__version__ = "1.0.0a3"
 __title__ = "Member Audit"