+++ memberaudit/tests/test_helpers.py
@@ -1,13 +1,14 @@
 from unittest.mock import Mock, patch
 
-
-from allianceauth.eveonline.evelinks import dotlan, evewho
-
-
+from django.contrib.auth.models import Group
 from django.test import TestCase
 
+from allianceauth.eveonline.evelinks import dotlan, evewho
+from allianceauth.tests.auth_utils import AuthUtils
+
 from .testdata.esi_client_stub import load_test_data
-from ..helpers import eve_xml_to_html
+from .testdata.load_entities import load_entities
+from ..helpers import eve_xml_to_html, users_with_permission
 
 
 class TestHTMLConversion(TestCase):
@@ -66,3 +67,45 @@ class TestHTMLConversion(TestCase):
                 .get("body")
             )
             self.assertTrue(result.find(dotlan.solar_system_url("Polaris")) != -1)
+
+
+class TestUsersWithPermissionQS(TestCase):
+    @classmethod
+    def setUpClass(cls) -> None:
+        super().setUpClass()
+        load_entities()
+        cls.permission = AuthUtils.get_permission_by_name("memberaudit.basic_access")
+        cls.group, _ = Group.objects.get_or_create(name="Test Group")
+        AuthUtils.add_permissions_to_groups([cls.permission], [cls.group])
+        cls.state = AuthUtils.create_state(name="Test State", priority=75)
+        cls.state.permissions.add(cls.permission)
+
+    def setUp(self) -> None:
+        self.user_1 = AuthUtils.create_user("Bruce Wayne")
+        self.user_2 = AuthUtils.create_user("Lex Luther")
+
+    @classmethod
+    def user_with_permission_pks(cls) -> set:
+        return set(users_with_permission(cls.permission).values_list("pk", flat=True))
+
+    def test_user_permission(self):
+        """direct user permissions"""
+        AuthUtils.add_permissions_to_user([self.permission], self.user_1)
+        self.assertSetEqual(self.user_with_permission_pks(), {self.user_1.pk})
+
+    def test_group_permission(self):
+        """group permissions"""
+        self.user_1.groups.add(self.group)
+        self.assertSetEqual(self.user_with_permission_pks(), {self.user_1.pk})
+
+    def test_state_permission(self):
+        """state permissions"""
+        AuthUtils.assign_state(self.user_1, self.state, disconnect_signals=True)
+        self.assertSetEqual(self.user_with_permission_pks(), {self.user_1.pk})
+
+    def test_distinct_qs(self):
+        """only return one user object, despiste multiple matches"""
+        AuthUtils.add_permissions_to_user([self.permission], self.user_1)
+        self.user_1.groups.add(self.group)
+        AuthUtils.assign_state(self.user_1, self.state, disconnect_signals=True)
+        self.assertSetEqual(self.user_with_permission_pks(), {self.user_1.pk})